refactor FamilyDetail
