+++ src/components/DashboardCard/DashboardCardField.tsx
@@ -3,17 +3,17 @@ import type { ReactNode } from 'react';
 
 interface Props {
   title: string;
-  value: ReactNode;
+  children: ReactNode;
 }
 
-export const DashboardCardField = ({ title, value }: Props) => {
+export const DashboardCardField = ({ title, children }: Props) => {
   return (
     <Box sx={{ textAlign: 'left' }}>
       <Title order={6} color="fgSubtle" weight={500}>
         {title}:
       </Title>
       <Space h={5} />
-      {value}
+      {children}
     </Box>
   );
 };
