+++ src/components/index.ts
@@ -1,6 +1,7 @@
 export * from './atoms';
 export * from './CreateFamily';
 export * from './CreateProject';
+export * from './DashboardCard';
 export * from './Fields';
 export * from './molecules';
 export * from './organisms';
