+++ src/pages/index.ts
@@ -1,4 +1,4 @@
-export * from './Dashboard/Card';
+export * from '../components/DashboardCard';
 export * from './Dashboard/Families';
 export * from './Dashboard/Projects';
 export * from './Login';
